refactor: extract core:auth interfaces to commonMain
Move AuthState.kt and AuthRepository interface to commonMain
source set, keeping AuthRepositoryImpl (Firebase/WebAuthn interop)
in wasmJsMain. Split build.gradle.kts dependencies accordingly.
