+++ core/auth/build.gradle.kts
@@ -10,11 +10,12 @@ kotlin {
     }
 
     sourceSets {
-        wasmJsMain.dependencies {
+        commonMain.dependencies {
             // compose.runtime が coroutines を推移的に提供
             api(compose.runtime)
             api(project(":shared"))
-
+        }
+        wasmJsMain.dependencies {
             implementation(libs.koin.core)
         }
     }
