+++ core/auth/src/wasmJsMain/kotlin/core/auth/AuthRepositoryImpl.kt
@@ -3,28 +3,6 @@ package core.auth
 import kotlinx.coroutines.await
 import model.User
 
-interface AuthRepository {
-    fun startListening()
-
-    suspend fun signIn(
-        email: String,
-        password: String,
-    ): Result<Unit>
-
-    suspend fun signOut(): Result<Unit>
-
-    suspend fun changePassword(
-        currentPassword: String,
-        newPassword: String,
-    ): Result<Unit>
-
-    suspend fun refreshToken(): String?
-
-    suspend fun signInWithCustomToken(token: String): Result<Unit>
-
-    fun isWebAuthnSupported(): Boolean
-}
-
 @OptIn(ExperimentalWasmJsInterop::class)
 class AuthRepositoryImpl : AuthRepository {
     private val auth by lazy { firebaseAuth(getFirebase()) }
